A few simple operations to perform on trees
Changed the has-children? method so that a node with a single child is still counted as having a child (slight bug). 

Changed from the left-child and right-child methods to n1, n2 and n3 child methods. This is a temporary solution, as nodes with two children will have to use the 'middle' (or n2-child method) to actually get the value of the node that is to the right. The reason for this is so that a node with more than two children, (such as T/X) can retrieve values from every child. 

Instead of thinking of the tree structure as [Parent left-child right-child], think of it as [Parent 0 1 2 ...].